Move orthogonal code out of proc.rs
