+++ smtlib/src/proc.rs
@@ -7,14 +7,13 @@
 //! parts are captured by the [`SolverCmd`] passed to launch the solver and in
 //! the code that parses models returned by [`SmtProc::check_sat`].
 
+use crate::conf::SolverCmd;
 use crate::sexp;
+use crate::tee::Tee;
 use nix::{errno::Errno, sys::signal, unistd::Pid};
-use std::hash::{Hash, Hasher};
 use std::process::ExitStatus;
 use std::{
-    collections::hash_map::DefaultHasher,
     ffi::{OsStr, OsString},
-    fs::OpenOptions,
     io::{self, BufRead, BufReader, ErrorKind, Write},
     path::{Path, PathBuf},
     process::{Child, ChildStdin, ChildStdout, Command, Stdio},
@@ -25,63 +24,6 @@ use thiserror::Error;
 use std::os::unix::process::ExitStatusExt;
 
 use super::sexp::{app, atom_s, sexp_l, Sexp};
-
-#[derive(Debug)]
-struct Tee {
-    dir: PathBuf,
-    contents: Vec<Sexp>,
-}
-
-fn calculate_hash<T: Hash>(v: T) -> String {
-    let mut hash_state = DefaultHasher::new();
-    v.hash(&mut hash_state);
-    let h = hash_state.finish();
-    return format!("{:016x}", h)[..8].to_string();
-}
-
-impl Tee {
-    fn new<P: AsRef<Path>>(dir: P) -> Self {
-        Self {
-            dir: dir.as_ref().to_path_buf(),
-            contents: vec![],
-        }
-    }
-
-    fn append(&mut self, s: Sexp) {
-        self.contents.push(s)
-    }
-
-    /// Save the SMT2 input currently sent to the solver to a file based on
-    /// content hash. Returns the saved file name.
-    fn save(&self) -> io::Result<PathBuf> {
-        let contents = self
-            .contents
-            .iter()
-            .map(|s| {
-                if let Sexp::Comment(c) = s {
-                    #[allow(clippy::comparison_to_empty)]
-                    if c == "" {
-                        return "".to_string();
-                    }
-                    return format!(";; {c}");
-                }
-                // TODO: this should be pretty-printed
-                s.to_string()
-            })
-            .collect::<Vec<_>>()
-            .join("\n");
-        let hash = calculate_hash(&contents);
-        let fname = PathBuf::from(format!("query-{hash}.smt2"));
-        let dest = self.dir.join(&fname);
-        let mut f = OpenOptions::new()
-            .create(true)
-            .write(true)
-            .truncate(true)
-            .open(dest)?;
-        write!(&mut f, "{contents}")?;
-        Ok(fname)
-    }
-}
 
 /// SmtProc wraps an instance of a solver process.
 #[derive(Debug)]
@@ -137,163 +79,6 @@ pub enum SolverError {
 }
 
 type Result<T> = std::result::Result<T, SolverError>;
-
-/// The full invocation of a solver binary.
-#[derive(Debug, Clone)]
-pub struct SolverCmd {
-    /// Binary to launch
-    pub cmd: String,
-    /// Arguments to pass
-    pub args: Vec<String>,
-    /// SMT options to send on startup
-    pub options: Vec<(String, String)>,
-}
-
-impl SolverCmd {
-    fn args<I, S>(&mut self, args: I)
-    where
-        I: IntoIterator<Item = S>,
-        S: AsRef<str>,
-    {
-        self.args
-            .extend(args.into_iter().map(|s| s.as_ref().to_string()));
-    }
-
-    /// Set an option.
-    pub fn option<S: AsRef<str>>(&mut self, name: &str, val: S) {
-        self.options
-            .push((name.to_string(), val.as_ref().to_string()));
-    }
-
-    fn cmdline(&self) -> String {
-        #[allow(clippy::useless_format)]
-        let args: Vec<_> = self
-            .args
-            .iter()
-            .map(|a| {
-                if a.contains(' ') {
-                    format!("\"{a}\"")
-                } else {
-                    format!("{a}")
-                }
-            })
-            .collect();
-        format!("{} {}", &self.cmd, args.join(" "))
-    }
-}
-
-/// Builder for creating a Z3 [`SolverCmd`].
-#[derive(Debug, Clone)]
-pub struct Z3Conf(SolverCmd);
-
-impl Z3Conf {
-    /// Create a Z3Conf with some default options. Uses `cmd` as the path to Z3.
-    pub fn new(cmd: &str) -> Self {
-        let mut cmd = SolverCmd {
-            cmd: cmd.to_string(),
-            args: vec![],
-            options: vec![],
-        };
-        cmd.args(["-in", "-smt2"]);
-        cmd.option("model.completion", "true");
-        let mut conf = Self(cmd);
-        conf.timeout_ms(Some(30000 * 100));
-        conf
-    }
-
-    /// Enable model compaction
-    pub fn model_compact(&mut self) {
-        self.0.option("model.compact", "true");
-    }
-
-    /// Set the SMT timeout option
-    pub fn timeout_ms(&mut self, ms: Option<usize>) {
-        // this is the default Z3 timeout
-        let ms = ms.unwrap_or(4294967295);
-        self.0.option("timeout", format!("{ms}"));
-    }
-
-    /// Get access to the raw options of the solver.
-    pub fn options(&mut self) -> &mut SolverCmd {
-        &mut self.0
-    }
-
-    /// Get the final command to run the solver.
-    pub fn done(self) -> SolverCmd {
-        self.0
-    }
-}
-
-/// Builder for a CVC4 or CVC5 [`SolverCmd`].
-#[derive(Debug, Clone)]
-pub struct CvcConf {
-    version5: bool,
-    cmd: SolverCmd,
-}
-
-impl CvcConf {
-    fn new_cvc(cmd: &str, version5: bool) -> Self {
-        let mut cmd = SolverCmd {
-            cmd: cmd.to_string(),
-            args: vec![],
-            options: vec![],
-        };
-        // for CVC4, --lang smt2 is needed when using stdin, but when run on a
-        // file with a .smt2 extension it will automatically use the right input
-        // format.
-        cmd.args(vec!["-q", "--lang", "smt2"]);
-        cmd.option("interactive", "false");
-        cmd.option("incremental", "true");
-        cmd.option("seed", "1");
-        Self { version5, cmd }
-    }
-
-    /// Create a new CVC4 builder with some default options.
-    pub fn new_cvc4(cmd: &str) -> Self {
-        Self::new_cvc(cmd, /*version5*/ false)
-    }
-
-    /// Create a new CVC5 builder with some default options.
-    pub fn new_cvc5(cmd: &str) -> Self {
-        Self::new_cvc(cmd, /*version5*/ true)
-    }
-
-    /// Enable finite model finding with mbqi.
-    pub fn finite_models(&mut self) {
-        self.cmd.option("finite-model-find", "true");
-        if self.version5 {
-            self.cmd.option("mbqi", "true");
-            self.cmd.option("fmf-mbqi", "fmc")
-        } else {
-            self.cmd.option("mbqi", "fmc");
-        }
-    }
-
-    /// Enable interleaving enumerative instantiation with other techniques.
-    pub fn interleave_enumerative_instantiation(&mut self) {
-        if self.version5 {
-            self.cmd.option("enum-inst-interleave", "true");
-        } else {
-            self.cmd.option("fs-interleave", "true");
-        }
-    }
-
-    /// Set a per-query time limit. None sets no time limit.
-    pub fn timeout_ms(&mut self, ms: Option<usize>) {
-        let ms = ms.unwrap_or(0);
-        self.cmd.option("tlimit-per", format!("{ms}"));
-    }
-
-    /// Get access to the raw options of the solver.
-    pub fn options(&mut self) -> &mut SolverCmd {
-        &mut self.cmd
-    }
-
-    /// Get the final command to run the solver.
-    pub fn done(self) -> SolverCmd {
-        self.cmd
-    }
-}
 
 impl Drop for SmtProc {
     fn drop(&mut self) {
@@ -645,8 +430,9 @@ impl SmtPid {
 #[cfg(test)]
 mod tests {
     use crate::{
+        conf::{CvcConf, Z3Conf},
         path::solver_path,
-        proc::{CvcConf, SatResp, SmtProc, SolverError, Z3Conf},
+        proc::{SatResp, SmtProc, SolverError},
         sexp::{app, atom_s, parse},
     };
     use eyre::Context;
